hogql base refactor
